termionix_client

Enum AnsiControlCode

pub enum AnsiControlCode {

    NUL,
    SOH,
    STX,
    ETX,
    EOT,
    ENQ,
    ACK,
    BEL,
    BS,
    HT,
    LF,
    VT,
    FF,
    CR,
    SO,
    SI,
    DLE,
    DC1,
    DC2,
    DC3,
    DC4,
    NAK,
    SYN,
    ETB,
    CAN,
    EM,
    SUB,
    FS,
    GS,
    RS,
    US,
    DEL,
    PAD,
    HOP,
    BPH,
    NBH,
    IND,
    NEL,
    SSA,
    ESA,
    HTS,
    HTJ,
    VTS,
    PLD,
    PLU,
    RI,
    SS2,
    SS3,
    DCS,
    PU1,
    PU2,
    STS,
    CCH,
    MW,
    SPA,
    EPA,
    SOS,
    SGCI,
    SCI,
    CSI,
    StC1,
    OscC1,
    PmC1,
    ApcC1,

}

ISO 6429 Control Codes (C0 and C1 sets)

Variants

NUL

Null character

SOH

Start of Heading

STX

Start of Text

ETX

End of Text

EOT

End of Transmission

ENQ

Enquiry

ACK

Acknowledge

BEL

Bell/Alert

BS

Backspace

HT

Horizontal Tab

LF

Line Feed

VT

Vertical Tab

FF

Form Feed

CR

Carriage Return

SO

Shift Out

SI

Shift In

DLE

Data Link Escape

DC1

Device Control 1

DC2

Device Control 2

DC3

Device Control 3

DC4

Device Control 4

NAK

Negatively Acknowledge

SYN

Synchronous Idle

ETB

End of Transmission Block

CAN

Cancel

EM

End of Medium

SUB

Substitute

FS

File Separator

GS

Group Separator

RS

Record Separator

US

Unit Separator

DEL

Delete

PAD

Padding Character

HOP

High Octet Preset

BPH

Break Permitted Here

NBH

No Break Here

IND

Index

NEL

Next Line

SSA

Start of Selected Area

ESA

End of Selected Area

HTS

Character Tabulation Set

HTJ

Character Tabulation with Justification

VTS

Line Tabulation Set

PLD

Partial Line Forward

PLU

Partial Line Backward

RI

Reverse Index

SS2

Single Shift Two

SS3

Single Shift Three

DCS

Device Control String

PU1

Private Use One

PU2

Private Use Two

STS

Set Transmit State

CCH

Cancel Character

MW

Message Waiting

SPA

Start of Guarded Area

EPA

End of Guarded Area

SOS

Start of String

SGCI

Single Graphic Character Introducer

SCI

Single Character Introducer

CSI

Control Sequence Introducer

StC1

String Terminator

OscC1

Operating System Command

PmC1

Privacy Message

ApcC1

Application Program Command

Implementations

impl AnsiControlCode

pub fn len(&self) -> usize

Returns the byte length of a single control code.

Control codes are always single bytes, so this method always returns 1. This method exists for consistency with other ANSI sequence types that have variable lengths.

Returns

Always returns 1, as all control codes (C0, C1, and DEL) are represented by a single byte.

Performance

This is an O(1) operation that simply returns a constant.

Examples

use termionix_ansicodec::ansi::AnsiControlCode;

assert_eq!(AnsiControlCode::NUL.len(), 1);
assert_eq!(AnsiControlCode::LF.len(), 1);
assert_eq!(AnsiControlCode::BEL.len(), 1);
assert_eq!(AnsiControlCode::DEL.len(), 1);

pub fn encode<T>(&self, dst: &mut T) -> Result<usize, AnsiCodecError>
where T: BufMut,

Encode this control code to a BufMut buffer.

This method encodes the control code as a single byte and writes it to the provided mutable buffer. The buffer’s write position is automatically advanced.

Arguments

dst - A mutable buffer implementing BufMut

Returns

Returns Ok(1) on success (exactly 1 byte is written). Returns an AnsiResult error if the buffer write fails.

Performance

This is an O(1) operation that writes a single byte.

Examples

use bytes::BytesMut;
use termionix_ansicodec::ansi::AnsiControlCode;

let code = AnsiControlCode::BEL;
let mut buffer = BytesMut::new();
let written = code.encode(&mut buffer).unwrap();
assert_eq!(written, 1);
assert_eq!(&buffer[..], b"\x07");

See Also

write() - Write to a std::io::Write trait object

pub fn write<W>(&self, writer: &mut W) -> Result<usize, Error>
where W: Write,

Write this control code to a std::io::Write writer.

This method performs the actual encoding of the control code into its single byte representation and writes it to the provided writer.

Arguments

writer - A mutable writer implementing std::io::Write

Returns

Returns Ok(1) on success (exactly 1 byte written). Returns std::io::Error if writing fails.

Encoding

Converts the control code to its byte representation using to_byte() and writes the single byte to the writer.

Examples

use termionix_ansicodec::ansi::AnsiControlCode;

let code = AnsiControlCode::LF;
let mut output = Vec::new();
let written = code.write(&mut output).unwrap();
assert_eq!(written, 1);
assert_eq!(output, b"\n");

See Also

encode() - Encode to a BufMut buffer

to_byte() - Get the byte value

pub fn to_byte(&self) -> u8

Convert control code to byte representation

pub fn from_byte(byte: u8) -> Option<AnsiControlCode>

Convert a byte to its corresponding control code