termionix_client

Trait TerminalHandler

pub trait TerminalHandler:
    Send
    + Sync
    + 'static {
    // Provided methods
    fn on_connect<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_disconnect<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_character<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _ch: char,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_line<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _line: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn on_bell<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_resize<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _width: usize,
        _height: usize,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_option_changed<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _option: TelnetOption,
        _enabled: bool,
        _local: bool,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_subnegotiation<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _subneg: TelnetArgument,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_error<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _error: ClientError,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_reconnect_attempt<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
        _attempt: u32,
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_reconnect_failed<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 TerminalConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Terminal client handler trait

Provided Methods

fn on_connect<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_disconnect<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_character<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _ch: char, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_line<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _line: &'life2 str, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn on_bell<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_resize<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _width: usize, _height: usize, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_option_changed<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _option: TelnetOption, _enabled: bool, _local: bool, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a Telnet option state changes

This is called when an option negotiation completes successfully, whether the option is being enabled or disabled.

Arguments

conn - The connection handle

option - The Telnet option that changed

enabled - true if the option was enabled, false if disabled

local - true if the option changed locally, false if remotely

fn on_subnegotiation<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _subneg: TelnetArgument, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a Telnet subnegotiation is received

This is called when a complete subnegotiation sequence is received from the server. Subnegotiations provide additional parameters for negotiated options.

fn on_error<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _error: ClientError, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_reconnect_attempt<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, _attempt: u32, ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn on_reconnect_failed<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 TerminalConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,